More compatibility

+++ IO/Integers.h
@@ -7,10 +7,10 @@ char MetricPrefix(int magnitude_base_1000) noexcept(false);
 
 [[nodiscard]] std::size_t ParseBytes(const std::string& bytes) noexcept(false);
 
-[[nodiscard]] constexpr int64 operator""_kB(uint64 v) noexcept { return v * 1024; }
-[[nodiscard]] constexpr int64 operator""_MB(uint64 v) noexcept { return v * 1024 * 1024; }
-[[nodiscard]] constexpr int64 operator""_GB(uint64 v) noexcept { return v * 1024 * 1024 * 1024; }
-[[nodiscard]] constexpr int64 operator""_TB(uint64 v) noexcept { return v * 1024 * 1024 * 1024 * 1024; }
-[[nodiscard]] constexpr int64 operator""_EB(uint64 v) noexcept { return v * 1024 * 1024 * 1024 * 1024 * 1024; }
-[[nodiscard]] constexpr int64 operator""_ZB(uint64 v) noexcept { return v * 1024 * 1024 * 1024 * 1024 * 1024 * 1024; }
-[[nodiscard]] constexpr int64 operator""_YB(uint64 v) noexcept { return v * 1024 * 1024 * 1024 * 1024 * 1024 * 1024 * 1024; }
+[[nodiscard]] constexpr int64 operator""_kB(unsigned long long v) noexcept { return v * 1024; }
+[[nodiscard]] constexpr int64 operator""_MB(unsigned long long v) noexcept { return v * 1024 * 1024; }
+[[nodiscard]] constexpr int64 operator""_GB(unsigned long long v) noexcept { return v * 1024 * 1024 * 1024; }
+[[nodiscard]] constexpr int64 operator""_TB(unsigned long long v) noexcept { return v * 1024 * 1024 * 1024 * 1024; }
+[[nodiscard]] constexpr int64 operator""_EB(unsigned long long v) noexcept { return v * 1024 * 1024 * 1024 * 1024 * 1024; }
+[[nodiscard]] constexpr int64 operator""_ZB(unsigned long long v) noexcept { return v * 1024 * 1024 * 1024 * 1024 * 1024 * 1024; }
+[[nodiscard]] constexpr int64 operator""_YB(unsigned long long v) noexcept { return v * 1024 * 1024 * 1024 * 1024 * 1024 * 1024 * 1024; }
